+++ lib/Echidna/Transaction.hs
@@ -25,7 +25,7 @@ import EVM.Types hiding (Env, VMOpts(timestamp, gasprice))
 import Echidna.ABI
 import Echidna.Orphans.JSON ()
 import Echidna.SourceMapping (lookupUsingCodehash)
-import Echidna.Symbolic (forceWord, forceAddr)
+import Echidna.SymExec.Symbolic (forceWord, forceAddr)
 import Echidna.Types (fromEVM)
 import Echidna.Types.Config (Env(..), EConfig(..))
 import Echidna.Types.Random
